requirements.txt
Flask==2.3.3
google-generativeai==0.8.3
python-dotenv==1.0.1